+++ lib/index.spec.ts
@@ -11,8 +11,14 @@ import { render } from './index';
 
 import completeResume from '@ianprime0509/jsonresume-schema/examples/valid/complete.json';
 
+import invalidEmailResume from '@ianprime0509/jsonresume-schema/examples/invalid/invalid-email.json';
+
 describe('render', () => {
-  it('renders a valid resume', () => {
+  it('renders a complete resume', () => {
     expect(render(completeResume)).to.be.a('string').that.is.not.empty;
+  });
+
+  it('throws an error when given an invalid resume', () => {
+    expect(() => render(invalidEmailResume)).to.throw(/invalid/i);
   });
 });